refactor: share api surface script option parsing
